Changelog: 1. Changed Course_elearnDAO to take double for learning hours instead of int for all methods to create or update course 2. Changed LoginDAO, UserDAO and User constructor to add address and contact number. added getter and setter for user.cs also 3. Masterpage changed URL path for /login.aspx to login.aspx 4. accountSetting.aspx updated with backend to auto fill and update contact and address. 5. applyCourse.aspx changed the URL from /Login.aspx to Login.aspx

+++ app/LearnHub/LearnHub/AppCode/dao/Course_elearnDAO.cs
@@ -251,7 +251,7 @@ namespace LearnHub.AppCode.dao
                         toReturn.setPrerequisite(list); //retrieve arraylist of all prereq course_elearn objects
                     }
                     toReturn.setCategoryID((int)dr["categoryID"]);//7
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
@@ -311,7 +311,7 @@ namespace LearnHub.AppCode.dao
                         toReturn.setPrerequisite(list); //retrieve arraylist of all prereq course_elearn objects
                     }
                     toReturn.setCategoryID((int)dr["categoryID"]);//7
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
@@ -366,7 +366,7 @@ namespace LearnHub.AppCode.dao
                     toReturn.setStatus((string)dr["status"]);
                     toReturn.setDescription((string)dr["description"]);
                     toReturn.setCategoryID((int)dr["categoryID"]);
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
@@ -427,7 +427,7 @@ namespace LearnHub.AppCode.dao
                     }
                     toReturn.setCategoryID((int)dr["categoryID"]);//7
                     toReturn.setCourseCreator(userDAO.getUserByID((string)dr["courseCreator"]));
-                    toReturn.setHoursAwarded((int)dr["hoursAwarded"]);
+                    toReturn.setHoursAwarded((double)dr["hoursAwarded"]);
                     if (!dr.IsDBNull(11))
                     {
                         toReturn.setTargetAudience((string)dr["targetAudience"]);
